requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
cffi==1.16.0
cryptography==42.0.8
Django==4.2.11
django-js-asset==2.0.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
jwt==1.3.1
packaging==23.2
Pillow==9.5.0
pycparser==2.22
pytz==2023.3.post1
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
whitenoise==6.6.0
